Finally working | Added a CLI
A program has been added which simply parses the given file (as the
first argument) and prints it out. This has been used to verify that the
library works.

The DUB config has been rewritten to move the library to ':lib' and
added the app under ':app'.

The Attribute type has been removed in favor of a value-string
associative array.

The Node type now has a well-formatted 'toString'-ifier. Unfortunately,
even if _we_ know that the inputs to the formatter is well-formed, it
still wants to throw errors, so a bunch of `assumeWontThrow` layers have
been added.

The Event type has also been removed as now the parser directly follows
the ParseTree from Pegged.

Additionally, a "standard" file to parse has been added, made of all the
code samples on the SDLang website (https://sdlang.org).